implement basic cli package
